Mostly hook together scrape and web
